requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.6.0
numpy==1.17.3
pandas==0.25.2
pytz==2019.3
pyarrow==0.15.0
pysolr==3.8.1